+++ tests/app/server.js
@@ -1988,11 +1988,19 @@ app.get('/svelte/frame-remember-skip', (req, res) =>
   inertia.render(req, res, { component: 'Svelte/FrameRememberSkip', props: {} }),
 )
 
+app.get('/svelte/frame-remember/toggle', (req, res) =>
+  inertia.render(req, res, { component: 'Svelte/FrameRememberToggle', props: {} }),
+)
+
 app.get('/svelte/frame-remember/pane', (req, res) =>
   inertia.render(req, res, { component: 'Svelte/FrameRememberPane', props: {} }),
 )
 
 app.get('/svelte/frame-remember/pane-skip', (req, res) =>
+  inertia.render(req, res, { component: 'Svelte/FrameRememberPane', props: {} }),
+)
+
+app.get('/svelte/frame-remember/pane-toggle', (req, res) =>
   inertia.render(req, res, { component: 'Svelte/FrameRememberPane', props: {} }),
 )